Refactor messages map
